+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
-## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [3.1.1]
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [3.2.0]
+This version adds session part properties to the ALFPath class and turns off save on delete by default.
+
+### Added
+
+- session part properties in one.alf.path.PureALFPath (lab, subject, date, sequence, collection, revision)
+- with_\* methods in one.alf.path.PureALFPath for replacing session parts
+
+### Modified
+
+- by default, cache tables are not saved when One.\_\_del\_\_ is called
+
+## [3.1.1]
 This version fixes issues with saving cache files.
 
 ### Modified